Muffles RuboCop warnings, while leaving roadmap for future work.
 RuboCop is a virtually industry-standard tool for verifying that Ruby
source files comply with community-originated Ruby language-usage
conventions. These conventions have been arrived at over the years as
an aid in improving the overall quality of maintained code.

However, introducing a tool such as RuboCop to an existing, legacy code
base commonly yields numerous violation messages, presenting an
apparently daunting task to the legacy maintainers. RuboCop has an
established convention for dealing with such situations; the standard
`.rubocop.yml` file is created with the single line
`inherit_from: .rubocop_todo.yml`. That latter file, produced by running
`rubocop --auto-gen-config`, is the result of analysing the existing code
and, for each "cop" or convention test known to RuboCop, explicitly
preventing files found to be in violation from raising violation notices.
Thus, the *existing* code base gets an apparent "clean bill of health";
RuboCop may then be included in the default Rake task flow without
problems. As files are modified going forward, the `.rubocop_todo.yml`
file should be consulted for mentions of that file. As developer time
permits, mentions of that source file in the `todo` config should be
individually removed, and the resulting violations cleaned up.

In that manner, an existing code base can gradually be brought into
compliance with conventions that many, if not most, Ruby OSS developers
are familiar with. This has the proven effect of reducing defect count
and developer workload as the "cleanliness" of the initially typically
wildly non-compliant codebase is brought up to snuff. All the while,
this does not materially add to the overall work required of the core
maintainers, as existing code "passes" inspection and any new changes
which introduce questionable code are called out immediately when the
code is still fresh in the mind of the developer who wrote that new code.

The result is a typically extremely significant quality win for the
maintainer team, and a reassurance to project teams considering the use
of the subject code (such as this Gem) for their own projects. What's not
to like?

As an example of incremental repair, the method
`Wice::Columns#yield_declaration` has had a guard clause introduced,
returning immediately if the value returned by calling
`#yield_declaration_of_column_filter` is `nil` and eliminating the need
for a previous all-encompassing `if` block. The file has correspondingly
been removed from the `Style/GuardClause` exclusion list in
`.rubocop_todo.yml`. Total time to make the change and verify that the
spec suite still gave a green bar: less than a minute.