Init wice grid on turbolinks page load
Attach `initWiceGrid` callback to the `turbolinks:load` event instead of the
`turbolinks:render` event.

The `turbolinks:render` event only fires when the page is updated, so attaching
the `initWiceGrid` callback to this event means that on the initial load the
function will not be called.

The `turbolinks:load` event is fired after each page change, which includes the
initial page load.